# internal/contacts/adapters/http/openapi/contacts.v1.oas.yaml
openapi: 3.0.3
info:
  title: Contacts API
  description: Contacts api backend
  version: 1.0.0
  contact: 
    name: david terranova
    email: davidp.terranova@gmail.com
servers:
  - url: http://localhost:8080/v1
    description: Localhost
tags:
  - name: "contacts"
    description: "Contacts API"
paths:
  /contacts:
    get:
      operationId: getContacts
      tags:
        - contacts
      summary: List all contacts
      security:
        - basicAuth: []
      responses:
        "200":
          description: "List all contacts"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createContact
      tags:
        - contacts
      summary: Create a new contact
      security:
        - basicAuth: []
      requestBody:
        description: Contact object that needs to be added
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  example: "John"
                last_name:
                  type: string
                  example: "Doe"
                email:
                  type: string
                  format: email
                  example: "jdoe@contact.local"
                phone:
                  type: string
                  format: phone
                  example: "+15555555555"
      responses:
        "201":
          description: "Create a new contact"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /contacts/{contactId}:
    put:
      operationId: updateContact
      tags:
        - contacts
      summary: Update an existing contact
      security:
        - basicAuth: []
      parameters:
        - $ref: "#/components/parameters/contactId"
      requestBody:
        description: Contact object that needs to be updated
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  example: "John"
                last_name:
                  type: string
                  example: "Doe"
                email:
                  type: string
                  format: email
                  example: "jdoe@contact.local"
                phone:
                  type: string
                  format: phone
                  example: "+15555555555"
      responses:
        "200":
          description: "Update an existing contact"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteContact
      tags:
        - contacts
      summary: Delete an existing contact
      security:
        - basicAuth: []
      parameters:
        - $ref: "#/components/parameters/contactId"
      responses:
        "204":
            description: "Delete an existing contact"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /contacts/{contactId}:export:
    get:
      operationId: exportContact
      tags:
        - contacts
      summary: Export an existing contact
      security:
        - basicAuth: []
      parameters:
        - $ref: "#/components/parameters/contactId"
      responses:
        "200":
          description: "Export an existing contact"
          content:
            text/vcard:
              schema:
                type: string
                format: string
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  parameters:
    contactId:
      in: path
      name: contactId
      description: "identifier of a contact"
      required: true
      schema:
        type: string
        format: uuid

  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
    Contact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        created_by:
          type: string
          format: uuid
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        email:
          type: string
          format: email
          example: "jdoe@contacts.local"
        phone:
          type: string
          format: phone
          example: "+15555555555"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
          description: "version number of the contact aggregate"
          example: 1

  securitySchemes:
    # bearerAuth:
    #   type: http
    #   scheme: bearer
    #   bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic